don't break MiningOperation, instead create new class
