+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/operations/MiningOperation.java
@@ -1,45 +1,32 @@
 package io.github.thebusybiscuit.slimefun4.implementation.operations;
 
-import java.util.OptionalInt;
-
 import javax.annotation.Nonnull;
 
 import com.google.common.base.Preconditions;
 
-import org.bukkit.block.Block;
 import org.bukkit.inventory.ItemStack;
 
-import io.github.thebusybiscuit.slimefun4.api.geo.GEOResource;
-import io.github.thebusybiscuit.slimefun4.api.geo.ResourceManager;
 import io.github.thebusybiscuit.slimefun4.core.machines.MachineOperation;
-import io.github.thebusybiscuit.slimefun4.implementation.Slimefun;
-import io.github.thebusybiscuit.slimefun4.implementation.items.geo.GEOMiner;
 
 /**
- * This {@link MachineOperation} represents a {@link GEOMiner}
- * mining a {@link GEOResource}.
+ * This {@link MachineOperation} represents an operation
+ * with no inputs, only a result.
  * 
  * @author TheBusyBiscuit
- * 
- * @see GEOMiner
  *
  */
 public class MiningOperation implements MachineOperation {
 
     private final ItemStack result;
 
-    private final GEOResource resource;
-    private final Block block;
     private final int totalTicks;
     private int currentTicks = 0;
 
-    public MiningOperation(@Nonnull GEOResource resource, Block block, int totalTicks) {
-        Preconditions.checkArgument(resource != null, "The resource cannot be null");
+    public MiningOperation(@Nonnull ItemStack result, int totalTicks) {
+        Preconditions.checkArgument(result != null, "The result cannot be null");
         Preconditions.checkArgument(totalTicks >= 0, "The amount of total ticks must be a positive integer or zero, received: " + totalTicks);
 
-        this.resource = resource;
-        this.result = resource.getItem().clone();
-        this.block = block;
+        this.result = result;
         this.totalTicks = totalTicks;
     }
 
@@ -67,13 +54,6 @@ public class MiningOperation implements MachineOperation {
     @Override
     public int getTotalTicks() {
         return totalTicks;
-    }
-
-    @Override
-    public void cancel() {
-        ResourceManager resourceManager = Slimefun.getGPSNetwork().getResourceManager();
-        OptionalInt supplies = resourceManager.getSupplies(resource, block.getWorld(), block.getX() >> 4, block.getZ() >> 4);
-        supplies.ifPresent(s -> resourceManager.setSupplies(resource, block.getWorld(), block.getX() >> 4, block.getZ() >> 4, s + 1));
     }
 
 }
